requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
Flask>=2.3.3
APScheduler>=3.10.4
setuptools>=69.0.0
Flask-Caching>=2.1.0 